fix: simplify contribution flow and production verification
