+++ .github/workflows/deploy.yml
@@ -244,7 +244,7 @@ jobs:
           retention-days: 7
 
       - name: Upload browser diagnostics
-        if: failure()
+        if: always()
         uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a
         with:
           name: slop-browser-diagnostics-${{ github.run_id }}
