+++ src/lib/leaderboard.ts
@@ -1,3 +1,80 @@
+import type {
+  AttributionAssessment,
+  AttributionAssessmentOptions,
+  AttributionCoverage,
+  EvidenceAssessment,
+  EvidenceCategory,
+  GitHubActor,
+  GitHubLabel,
+  GitHubTextSource,
+  InvalidAttributionMarker,
+  IssueRecord,
+  LeaderboardEntry,
+  LeaderboardInput,
+  LeaderboardMethodology,
+  LeaderboardSnapshot,
+  MergedPullRequestOutcome,
+  MergedPullRequestReviewRecord,
+  ModelAttribution,
+  PullRequestFile,
+  PullRequestRecord,
+  PullRequestReview,
+  ReviewExclusion,
+  ReviewExclusionReason,
+  ScoreCategory,
+  ScoreEvent,
+  ScoreOpportunity,
+  ScoreOpportunityKind,
+  VerifiedEvidenceArtifact,
+  WorkItem,
+  WorkItemCandidateExclusion,
+  WorkItemClaimStatus,
+  WorkItemEvidenceStatus,
+  WorkItemModelStatus,
+  WorkItemSelection,
+} from "./leaderboard-types";
+
+export type {
+  AttributionAssessment,
+  AttributionAssessmentOptions,
+  AttributionCoverage,
+  CapUsageBucket,
+  CapUsageStatus,
+  EvidenceAssessment,
+  EvidenceCategory,
+  EvidenceFinding,
+  GitHubActor,
+  GitHubActorKind,
+  GitHubLabel,
+  GitHubTextSource,
+  InvalidAttributionMarker,
+  IssueRecord,
+  LeaderboardEntry,
+  LeaderboardInput,
+  LeaderboardMethodology,
+  LeaderboardSnapshot,
+  LeaderboardSourceMetadata,
+  MergedPullRequestOutcome,
+  MergedPullRequestReviewRecord,
+  ModelAttribution,
+  PullRequestFile,
+  PullRequestRecord,
+  PullRequestReview,
+  ReviewExclusion,
+  ReviewExclusionReason,
+  ScoreCategory,
+  ScoreEvent,
+  ScoreOpportunity,
+  ScoreOpportunityKind,
+  VerifiedEvidenceArtifact,
+  WorkItem,
+  WorkItemCandidateExclusion,
+  WorkItemClaimStatus,
+  WorkItemEvidenceStatus,
+  WorkItemModelStatus,
+  WorkItemSelection,
+} from "./leaderboard-types";
+
 /**
  * Defines the public contribution snapshot and the deterministic scoring policy
  * for slop.cash. GitHub ingestion stays outside this module so fixtures can
@@ -17,7 +94,6 @@ import {
   PRIMARY_REPOSITORY,
   type RepositoryId,
   TARGET_REPOSITORIES,
-  type TargetRepository,
 } from "./repositories.mjs";
 import {
   assertReviewRecord,
@@ -83,21 +159,6 @@ export function mergedPullRequestPoints(ordinal: number): number {
   return Math.max(1, Math.ceil(10 / Math.sqrt(ordinal)));
 }
 
-export type GitHubActorKind =
-  | "Bot"
-  | "Mannequin"
-  | "Organization"
-  | "User"
-  | "Unknown";
-
-export interface GitHubActor {
-  id: string;
-  login: string;
-  avatarUrl: string;
-  url: string;
-  kind: GitHubActorKind;
-}
-
 // GitHub appends a mutable ?u= cache key to avatar URLs whenever a user
 // changes their avatar, so the same actor can carry different avatarUrl bytes
 // in live data and in reviewed manifests. Only the stable v parameter may
@@ -107,48 +168,6 @@ export function canonicalActorAvatarUrl(value: string): string {
   const version = url.searchParams.get("v");
   url.search = version === null ? "" : `v=${version}`;
   return url.toString();
-}
-
-export interface GitHubLabel {
-  id: string;
-  name: string;
-  color: string;
-}
-
-export interface GitHubTextSource {
-  id: string;
-  artifactId: string;
-  kind: "body" | "comment" | "review";
-  body: string;
-  url: string;
-  createdAt: string;
-  updatedAt: string;
-  author: GitHubActor | null;
-  authorAssociation?: string | null;
-  artifactUrl?: string;
-  artifactHeadSha?: string;
-}
-
-export interface PullRequestFile {
-  path: string;
-  additions: number;
-  deletions: number;
-}
-
-export interface PullRequestReview {
-  id: string;
-  body: string;
-  state:
-    | "APPROVED"
-    | "CHANGES_REQUESTED"
-    | "COMMENTED"
-    | "DISMISSED"
-    | "PENDING"
-    | string;
-  submittedAt: string | null;
-  url: string;
-  author: GitHubActor | null;
-  inlineCommentCount: number;
 }
 
 export const REVIEW_EXCLUSION_REASONS = [
@@ -167,492 +186,6 @@ export const REVIEW_EXCLUSION_REASONS = [
   "external-prize-policy",
   "reviewer-cycle-cap",
 ] as const;
-
-export type ReviewExclusionReason = (typeof REVIEW_EXCLUSION_REASONS)[number];
-
-export interface ReviewExclusion {
-  id: string;
-  reviewId: string;
-  pullRequestId: string;
-  pullRequestNumber: number;
-  repository: RepositoryId;
-  url: string;
-  reason: ReviewExclusionReason;
-}
-
-export interface PullRequestRecord {
-  id: string;
-  number: number;
-  title: string;
-  url: string;
-  body: string;
-  createdAt: string;
-  updatedAt: string;
-  lastEditedAt: string | null;
-  /**
-   * GitHub's last body editor. Used only to distinguish author post-merge
-   * evidence farming from third-party (usually maintainer) body touches that
-   * must not void a head-pinned pre-merge evidence package.
-   */
-  editor: GitHubActor | null;
-  mergedAt: string | null;
-  headRefOid: string;
-  isDraft: boolean;
-  reviewDecision: string | null;
-  activeReviewRequestCount: number;
-  author: GitHubActor | null;
-  assignees: GitHubActor[];
-  labels: GitHubLabel[];
-  files: PullRequestFile[];
-  comments: GitHubTextSource[];
-  reviews: PullRequestReview[];
-  closingIssueIds: string[];
-  additions: number;
-  deletions: number;
-  changedFiles: number;
-  commitCount: number;
-}
-
-/**
- * Complete text and review evidence for a merged pull request whose expensive
- * author-detail connections were not requested. Detail-only fields are
- * deliberately absent rather than represented by plausible empty values.
- */
-export type MergedPullRequestReviewRecord = Pick<
-  PullRequestRecord,
-  | "id"
-  | "number"
-  | "title"
-  | "url"
-  | "body"
-  | "createdAt"
-  | "updatedAt"
-  | "lastEditedAt"
-  | "editor"
-  | "mergedAt"
-  | "headRefOid"
-  | "author"
-  | "comments"
-  | "reviews"
->;
-
-export interface MergedPullRequestOutcome {
-  id: string;
-  number: number;
-  title: string;
-  url: string;
-  body: string;
-  createdAt: string;
-  updatedAt: string;
-  mergedAt: string;
-  author: GitHubActor | null;
-  additions: number;
-  deletions: number;
-}
-
-export interface IssueRecord {
-  id: string;
-  number: number;
-  title: string;
-  url: string;
-  body: string;
-  createdAt: string;
-  updatedAt: string;
-  closedAt: string | null;
-  stateReason: "COMPLETED" | "NOT_PLANNED" | "REOPENED" | null | string;
-  author: GitHubActor | null;
-  assignees: GitHubActor[];
-  labels: GitHubLabel[];
-  comments: GitHubTextSource[];
-  closedByPullRequests: Array<{
-    id: string;
-    number: number;
-    url: string;
-    mergedAt: string | null;
-    body: string;
-    createdAt: string;
-    updatedAt: string;
-    author: GitHubActor | null;
-  }>;
-}
-
-export type EvidenceCategory =
-  | "screenshot"
-  | "video"
-  | "logs"
-  | "trajectory"
-  | "domain-artifact";
-
-export interface EvidenceFinding {
-  category: EvidenceCategory;
-  points: number;
-  sourceIds: string[];
-}
-
-export interface EvidenceAssessment {
-  points: number;
-  maxPoints: 6;
-  categories: EvidenceCategory[];
-  findings: EvidenceFinding[];
-}
-
-/**
- * Records one artifact that the live generator fetched and structurally
- * verified. The source body is retained only in the in-memory generation input
- * so a verdict cannot be replayed after a PR body or comment is edited.
- */
-export interface VerifiedEvidenceArtifact {
-  pullRequestId: string;
-  pullRequestMergedAt: string | null;
-  pullRequestHeadOid: string;
-  pullRequestUpdatedAt: string;
-  sourceId: string;
-  sourceBody: string;
-  sourceUpdatedAt: string;
-  category: EvidenceCategory;
-  artifactIdentity: string;
-  contentSha256: string;
-}
-
-export interface InvalidAttributionMarker {
-  sourceId: string;
-  sourceUrl: string;
-  reason: string;
-}
-
-export interface ModelAttribution {
-  id: string;
-  sourceId: string;
-  sourceUrl: string;
-  artifactId: string;
-  actor: GitHubActor | null;
-  provider: string;
-  model: string;
-  identifier: string;
-  client: string | null;
-  skillRevision: string | null;
-  run: ProjectRunReceipt | null;
-  format: "machine-marker" | "visible-declaration";
-  status: "self-reported";
-}
-
-export interface AttributionAssessment {
-  declarations: ModelAttribution[];
-  invalidMarkers: InvalidAttributionMarker[];
-  coverage: AttributionCoverage;
-}
-
-export interface AttributionAssessmentOptions {
-  requireEverySource?: boolean;
-  verifyRunReceipt?: (value: unknown) => ProjectRunReceipt;
-}
-
-export interface AttributionCoverage {
-  status: "complete" | "partial" | "missing" | "invalid";
-  eligibleSourceCount: number;
-  validSourceCount: number;
-  missingSourceCount: number;
-  invalidSourceCount: number;
-  humanOnlySourceCount: number;
-}
-
-export type ScoreCategory =
-  | "merged-pull-request"
-  | "resolved-issue"
-  | "material-test-change"
-  | "evidence"
-  | "substantive-review"
-  | "evaluated-contribution";
-
-export interface ScoreEvent {
-  id: string;
-  actor: GitHubActor;
-  category: ScoreCategory;
-  points: number;
-  scoreThirds?: number;
-  evidenceBonusBasisPoints?: 0 | 1_000 | 1_500 | 2_500;
-  workUnitId?: string;
-  scoreDecisionSourceId?: string;
-  occurredAt: string;
-  repository: RepositoryId;
-  source: {
-    id: string;
-    kind: "comment" | "issue" | "pull-request" | "review";
-    number: number;
-    title: string;
-    url: string;
-  };
-  reason: string;
-  continuity?: {
-    sourceSnapshotSha256: string;
-    decisionUrl: string;
-  };
-  evaluation?: {
-    decisionUrl: string;
-    reviewedAt: string;
-    reviewer: string;
-    manifestPath: string;
-    manifestSha256: string;
-  };
-}
-
-export type ScoreOpportunityKind =
-  | "near-material-test"
-  | "expand-review"
-  | "missing-evidence"
-  | "partial-evidence";
-
-/** Still-actionable scoring opportunities for public profile guidance. */
-export interface ScoreOpportunity {
-  id: string;
-  actor: GitHubActor;
-  kind: ScoreOpportunityKind;
-  category: ScoreCategory;
-  /** Null when the row is useful guidance but cannot add standalone score. */
-  potentialPoints: number | null;
-  occurredAt: string;
-  repository: RepositoryId;
-  source: ScoreEvent["source"];
-  reason: string;
-  hint: string;
-}
-
-export interface CapUsageBucket {
-  used: number;
-  cap: number | null;
-}
-
-/** Per-contributor monthly cap fill for compact profile status lines. */
-export interface CapUsageStatus {
-  month: string;
-  mergedPullRequests: CapUsageBucket;
-  resolvedIssues: CapUsageBucket;
-  materialTestChanges: CapUsageBucket;
-  evidencePoints: CapUsageBucket;
-  substantiveReviews: CapUsageBucket;
-  evaluatedContributions: CapUsageBucket;
-}
-
-export interface LeaderboardEntry {
-  rank: number;
-  actor: GitHubActor;
-  score: number;
-  scoreThirds: number;
-  points: {
-    mergedPullRequests: number;
-    resolvedIssues: number;
-    materialTestChanges: number;
-    evidence: number;
-    substantiveReviews: number;
-    evaluatedContributions: number;
-  };
-  pointThirds: {
-    mergedPullRequests: number;
-    resolvedIssues: number;
-    materialTestChanges: number;
-    evidence: number;
-    substantiveReviews: number;
-    evaluatedContributions: number;
-  };
-  acceptedOutcomes: {
-    mergedPullRequests: number;
-    resolvedIssues: number;
-    materialTestChanges: number;
-    evidenceCategories: number;
-    substantiveReviews: number;
-    evaluatedContributions: number;
-  };
-  rawActivity: {
-    comments: number;
-    reviews: number;
-    commits: number;
-    additions: number;
-    deletions: number;
-  };
-  reportedModels: string[];
-}
-
-export interface WorkItemClaimStatus {
-  status: "claimed" | "unclaimed";
-  source: "assignee" | "label" | "claim-comment" | "none";
-  kind: "implementation" | "review" | null;
-  actors: GitHubActor[];
-  claimedAt: string | null;
-}
-
-export interface WorkItemEvidenceStatus {
-  status: "complete" | "partial" | "missing";
-  points: number;
-  maxPoints: 6;
-  categories: EvidenceCategory[];
-}
-
-export interface WorkItemModelStatus {
-  status: AttributionCoverage["status"];
-  identifiers: string[];
-  machineMarkerCount: number;
-  invalidMarkerCount: number;
-  eligibleSourceCount: number;
-  validSourceCount: number;
-  missingSourceCount: number;
-  invalidSourceCount: number;
-  humanOnlySourceCount: number;
-  provenance: "self-reported" | "none";
-}
-
-export type WorkItemCandidateExclusion =
-  | "active-review-request"
-  | "already-approved"
-  | "blocked"
-  | "bot-authored"
-  | "changes-requested"
-  | "claimed"
-  | "draft"
-  | "security-sensitive"
-  | "untriaged"
-  | "unknown-author";
-
-export interface WorkItemSelection {
-  status: "candidate" | "excluded";
-  reasons: WorkItemCandidateExclusion[];
-}
-
-export interface WorkItem {
-  id: string;
-  kind: "issue" | "pull-request";
-  number: number;
-  title: string;
-  url: string;
-  repository: RepositoryId;
-  author: GitHubActor | null;
-  createdAt: string;
-  updatedAt: string;
-  labels: string[];
-  priority: "urgent" | "high" | "normal" | "low";
-  actionability: "actionable" | "blocked" | "draft";
-  isDraft: boolean | null;
-  reviewDecision: string | null;
-  activeReviewRequestCount: number | null;
-  commentCount: number;
-  claim: WorkItemClaimStatus;
-  selection: WorkItemSelection;
-  evidence: WorkItemEvidenceStatus;
-  model: WorkItemModelStatus;
-}
-
-export interface LeaderboardMethodology {
-  summary: string;
-  scoringRules: Array<{
-    id: ScoreCategory;
-    points: string;
-    cap: string;
-    qualification: string;
-  }>;
-  evidenceWeights: Record<EvidenceCategory, number>;
-  materialTestThreshold: {
-    minimumAdditions: number;
-    minimumTotalChurn: number;
-    cap: string;
-  };
-  exclusions: string[];
-  nonScoringActivity: string[];
-  provenancePolicy: string;
-  collectionPolicy: string;
-}
-
-export interface LeaderboardSourceMetadata {
-  provider: "github-graphql";
-  fetchedAt: string;
-  cutoffAt: string;
-  repositoryId: string;
-  repositories: Array<{ id: RepositoryId; repositoryId: string }>;
-  requestCount: number;
-  searchSliceCount: number;
-  rateLimit: {
-    cost: number;
-    consumedDuringRun?: number;
-    limit: number;
-    remaining: number;
-    resetAt: string;
-  };
-  counts: {
-    mergedPullRequests: number;
-    detailedMergedPullRequests: number;
-    closedIssues: number;
-    detailedClosedIssues: number;
-    resolvedIssues: number;
-    openIssues: number;
-    openPullRequests: number;
-  };
-  verificationWindow: {
-    days: number;
-    from: string;
-    to: string;
-  };
-  evidenceVerification: {
-    status: "complete" | "suppressed-limit";
-    sourceCount: number;
-    artifactCount: number;
-    maxSources: number;
-    maxArtifacts: number;
-  };
-}
-
-export interface LeaderboardSnapshot {
-  schemaVersion: typeof LEADERBOARD_SCHEMA_VERSION;
-  repository: typeof LEADERBOARD_REPOSITORY;
-  repositories: Omit<TargetRepository, "aliases" | "expectedNodeId">[];
-  ruleVersion: typeof SCORE_RULE_VERSION;
-  generatedAt: string;
-  sourceUpdatedAt: string;
-  stale: false;
-  window: {
-    days: number;
-    from: string;
-    to: string;
-  };
-  methodology: LeaderboardMethodology;
-  source: LeaderboardSourceMetadata;
-  leaders: LeaderboardEntry[];
-  ledger: ScoreEvent[];
-  reviewExclusions?: ReviewExclusion[];
-  opportunities: ScoreOpportunity[];
-  attributions: ModelAttribution[];
-  invalidAttributionMarkers: InvalidAttributionMarker[];
-  attributionCoverage: AttributionCoverage;
-  workQueue: {
-    issues: WorkItem[];
-    pullRequests: WorkItem[];
-  };
-}
-
-export interface LeaderboardInput {
-  generatedAt: string;
-  windowFrom: string;
-  windowTo: string;
-  sourceUpdatedAt: string;
-  source: LeaderboardSourceMetadata;
-  mergedPullRequestOutcomes: MergedPullRequestOutcome[];
-  mergedPullRequests: PullRequestRecord[];
-  reviewedMergedPullRequests?: MergedPullRequestReviewRecord[];
-  detailEligibleMergedPullRequestIds: string[];
-  closedIssueCount: number;
-  resolvedIssues: IssueRecord[];
-  openIssues: IssueRecord[];
-  openPullRequests: PullRequestRecord[];
-  verificationWindowFrom: string;
-  verifiedEvidence: VerifiedEvidenceArtifact[];
-  evaluatedContributions?: ScoreEvent[];
-  /**
-   * Previously published accepted review events. GitHub may delete a merged
-   * pull request and its reviews; only events whose parent outcome is absent
-   * are replayed, so live GitHub remains authoritative whenever it exists.
-   */
-  retainedReviewEvents?: ScoreEvent[];
-  /** Signed attributions recovered from the same accepted snapshot. */
-  retainedReviewAttributions?: ModelAttribution[];
-  verifyRunReceipt?: (value: unknown) => ProjectRunReceipt;
-}
 
 function exactEvidenceBonusRun(
   event: ScoreEvent,
@@ -2740,18 +2273,18 @@ function collectOpenPullRequestOpportunities(
       // Drafts are still open, but they are not ready for score-facing guidance.
       if (!pullRequest.isDraft) {
         if (isNearMaterialTestChange(pullRequest.files)) {
-          const { additions, churn } = materialTestStats(pullRequest.files);
           opportunities.push({
             id: `${pullRequest.id}:opportunity:near-material-test`,
             actor: pullRequest.author,
             kind: "near-material-test",
             category: "material-test-change",
-            potentialPoints: 4,
+            potentialPoints: null,
             occurredAt: pullRequest.updatedAt,
             repository,
             source: pullRequestSource,
-            reason: `Recognized test files currently add ${additions} lines and change ${churn} total lines; thresholds are ${MATERIAL_TEST_ADDITIONS} additions and ${MATERIAL_TEST_CHURN} churn.`,
-            hint: `Add recognized test coverage to reach ${MATERIAL_TEST_ADDITIONS} additions and ${MATERIAL_TEST_CHURN} total churn before merge.`,
+            reason:
+              "Test changes should demonstrate the accepted behavior or a reproduced regression; line counts do not establish usefulness.",
+            hint: "Test the behavior changed by this pull request. Do not add tests or lines merely to increase a score.",
           });
         }
 
@@ -2843,7 +2376,7 @@ function collectOpenPullRequestOpportunities(
         actor: review.author,
         kind: "expand-review",
         category: "substantive-review",
-        potentialPoints: 3,
+        potentialPoints: null,
         occurredAt: submittedAt,
         repository,
         source: {
@@ -2854,8 +2387,8 @@ function collectOpenPullRequestOpportunities(
           url: review.url,
         },
         reason:
-          "Review is APPROVED or CHANGES_REQUESTED but still needs at least 20 characters of rationale or an inline comment.",
-        hint: "Add at least 20 characters of review rationale or an inline comment before merge.",
+          "The review does not yet explain its approval or requested changes.",
+        hint: "Explain a concrete correctness finding or why the change is ready. Padding the review does not make it useful.",
       });
     }
   }
@@ -3493,7 +3026,7 @@ export function createLeaderboardSnapshot(
             title: pullRequest.title,
             url: source.url,
           },
-          reason: `Automated ${record.reviewLoad} review with finalized private trace; maintainer scoring remains authoritative.`,
+          reason: `Automated ${record.reviewLoad} review${receipt.traceUpload ? " with finalized private trace" : " without private trace"}; maintainer scoring remains authoritative.`,
         })
       ) {
         recordScoredSources([source]);
@@ -5129,10 +4662,10 @@ function assertOpportunityValue(
   const validPair =
     (kind === "near-material-test" &&
       category === "material-test-change" &&
-      potentialPoints === 4) ||
+      (potentialPoints === null || potentialPoints === 4)) ||
     (kind === "expand-review" &&
       category === "substantive-review" &&
-      potentialPoints === 3) ||
+      (potentialPoints === null || potentialPoints === 3)) ||
     ((kind === "missing-evidence" || kind === "partial-evidence") &&
       category === "evidence" &&
       (potentialPoints === null ||
@@ -5150,7 +4683,9 @@ function assertOpportunityValue(
     opportunity.hint.length > 256 ||
     !(
       (kind === "near-material-test" &&
-        opportunity.hint.startsWith("Add recognized test coverage")) ||
+        (opportunity.hint.startsWith("Add recognized test coverage") ||
+          opportunity.hint ===
+            "Test the behavior changed by this pull request. Do not add tests or lines merely to increase a score.")) ||
       (kind === "missing-evidence" &&
         opportunity.hint ===
           "Add verified screenshot, video, or log evidence before merge.") ||
@@ -5158,8 +4693,10 @@ function assertOpportunityValue(
         opportunity.hint ===
           "Finish verified evidence categories before merge.") ||
       (kind === "expand-review" &&
-        opportunity.hint ===
-          "Add at least 20 characters of review rationale or an inline comment before merge.")
+        (opportunity.hint ===
+          "Add at least 20 characters of review rationale or an inline comment before merge." ||
+          opportunity.hint ===
+            "Explain a concrete correctness finding or why the change is ready. Padding the review does not make it useful."))
     )
   ) {
     throw new Error(`${path}.hint must be an actionable next step`);
