+++ src/lib/project-proposal.ts
@@ -1,0 +1,93 @@
+import {
+  formatMonthlyCapDisplay,
+  MAX_MONTHLY_CAP_MINOR,
+} from "./project-schema.mjs";
+import { PROJECTS, type ProjectDefinition } from "./projects.mjs";
+export function rootPublishedTemplateProject(
+  projects: readonly ProjectDefinition[] = PROJECTS,
+): ProjectDefinition {
+  const publishers = projects.filter(
+    (project) => project.skill.publishAtRoot === true,
+  );
+  if (publishers.length !== 1) {
+    throw new TypeError(
+      "project registry must declare exactly one root-published template skill",
+    );
+  }
+  return publishers[0];
+}
+
+export const ROOT_PUBLISHED_TEMPLATE = rootPublishedTemplateProject();
+
+export function safeProposalHttpsUrl(value: string): boolean {
+  if (value.length === 0 || value.length > 500) return false;
+  try {
+    const parsed = new URL(value);
+    return (
+      parsed.protocol === "https:" &&
+      parsed.hostname.length > 0 &&
+      !parsed.username &&
+      !parsed.password &&
+      !parsed.hash
+    );
+  } catch {
+    return false;
+  }
+}
+
+export function immutableProposalTermsUrl(
+  value: string,
+  repository: string,
+  commit: string,
+): boolean {
+  if (!safeProposalHttpsUrl(value)) return false;
+  const parsed = new URL(value);
+  const prefix = `/${repository}/blob/${commit}/`;
+  return (
+    parsed.origin === "https://github.com" &&
+    !parsed.search &&
+    parsed.pathname.startsWith(prefix) &&
+    parsed.pathname.length > prefix.length
+  );
+}
+
+export function boundedText(
+  value: string,
+  minimum: number,
+  maximum: number,
+): boolean {
+  const length = value.trim().length;
+  return length >= minimum && length <= maximum;
+}
+
+export function slugify(value: string): string {
+  return value
+    .toLowerCase()
+    .trim()
+    .replace(/[^a-z0-9]+/gu, "-")
+    .replace(/^-|-$/gu, "")
+    .slice(0, 48);
+}
+
+export function monthlyPoolValue(value: string): {
+  display: string;
+  minor: string;
+  valid: boolean;
+} {
+  if (!/^(?:0|[1-9]\d{0,9})(?:\.\d{1,2})?$/u.test(value)) {
+    return { display: "$0", minor: "0", valid: false };
+  }
+  const [whole, fraction = ""] = value.split(".");
+  const minor = (
+    BigInt(whole) * 1_000_000n +
+    BigInt(fraction.padEnd(2, "0")) * 10_000n
+  ).toString();
+  if (BigInt(minor) > MAX_MONTHLY_CAP_MINOR) {
+    return { display: "$0", minor: "0", valid: false };
+  }
+  return {
+    display: formatMonthlyCapDisplay(minor),
+    minor,
+    valid: true,
+  };
+}
