+++ src/lib/source-repository.ts
@@ -1,0 +1,1 @@
+export const SOURCE_REPOSITORY = "https://github.com/SlopDotCash/slopdotcash";
